+++ audio-api/src/main/resources/no/ndla/audioapi/db/migration/V1__Create_Initial_tables.sql
