+++ audio-api/src/main/resources/no/ndla/audioapi/db/migration/V3__AddOptimisticLocking.sql
