+++ audio-api/src/main/scala/no/ndla/audioapi/db/migration/V13__AddSeriesDescriptionField.scala
@@ -5,50 +5,45 @@
  * See LICENSE
  */
 
-package audioapi.db.migration
+package no.ndla.audioapi.db.migration
 
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
+import org.json4s.JObject
 import org.json4s.JsonAST.JField
-import org.json4s.ext.JavaTimeSerializers
 import org.json4s.native.JsonMethods.{compact, parse, render}
-import org.json4s.{DefaultFormats, Extraction, Formats, JObject}
 import org.postgresql.util.PGobject
 import scalikejdbc.{DB, DBSession, _}
 
-import java.time.LocalDateTime
-
-class V12__AddSeriesDateField extends BaseJavaMigration {
-  private implicit val formats: Formats = DefaultFormats ++ JavaTimeSerializers.all
-  private val dateToUse                 = LocalDateTime.now()
-  private val jsonDate                  = Extraction.decompose(dateToUse)(formats)
+class V13__AddSeriesDescriptionField extends BaseJavaMigration {
 
   override def migrate(context: Context): Unit = {
     val db = DB(context.getConnection)
     db.autoClose(false)
 
     db.withinTx { implicit session =>
-      allAudios.map { case (id: Long, document: String) =>
+      allSeries.map { case (id: Long, document: String) =>
         update(convertDocument(document), id)
       }
     }
   }
 
-  def allAudios(implicit session: DBSession): List[(Long, String)] = {
+  def allSeries(implicit session: DBSession): List[(Long, String)] = {
     sql"select id, document from seriesdata"
       .map(rs => (rs.long("id"), rs.string("document")))
       .list()
   }
 
   def convertDocument(document: String): String = {
-    val oldArticle = parse(document)
+    val oldSeries = parse(document)
 
-    val objectToMerge = JObject(
-      JField("created", jsonDate),
-      JField("updated", jsonDate)
-    )
+    val descriptions = (oldSeries \ "title").mapField {
+      case "title" -> x => "description" -> x
+      case x            => x
+    }
 
-    val newArticle = oldArticle.merge(objectToMerge)
-    compact(render(newArticle))
+    val objectToMerge = JObject(JField("description", descriptions))
+    val newSeries     = oldSeries.merge(objectToMerge)
+    compact(render(newSeries))
   }
 
   def update(document: String, id: Long)(implicit session: DBSession): Int = {
