+++ audio-api/src/main/scala/no/ndla/audioapi/db/migration/V2__AddUpdatedColoums.scala
@@ -5,7 +5,7 @@
  * See LICENSE
  */
 
-package audioapi.db.migration
+package no.ndla.audioapi.db.migration
 
 import org.flywaydb.core.api.migration.{BaseJavaMigration, Context}
 import org.json4s.native.JsonMethods._
