+++ audio-api/src/test/scala/no/ndla/audioapi/db/migration/V14__CreateMissingFilePathsTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package audioapi.db.migration
+package no.ndla.audioapi.db.migration
 
-import audioapi.db.migration.V14__CreateMissingFilePaths
+import no.ndla.audioapi.db.migration.V14__CreateMissingFilePaths
 import no.ndla.audioapi.{TestEnvironment, UnitSuite}
 
 class V14__CreateMissingFilePathsTest extends UnitSuite with TestEnvironment {
