+++ audio-api/src/test/scala/no/ndla/audioapi/db/migration/V15__ConvertLanguageUnknownTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package audioapi.db.migration
+package no.ndla.audioapi.db.migration
 
-import audioapi.db.migration.V15__ConvertLanguageUnknown
+import no.ndla.audioapi.db.migration.V15__ConvertLanguageUnknown
 import no.ndla.audioapi.{TestEnvironment, UnitSuite}
 
 class V15__ConvertLanguageUnknownTest extends UnitSuite with TestEnvironment {
