+++ audio-api/src/test/scala/no/ndla/audioapi/db/migration/V4__AddLanguageToAllTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package audioapi.db.migration
+package no.ndla.audioapi.db.migration
 
-import audioapi.db.migration.{V4_Audio, V4_AudioMetaInformation, V4_Copyright, V4_Tag, V4_Title, V4__AddLanguageToAll}
+import no.ndla.audioapi.db.migration.{V4_Audio, V4_AudioMetaInformation, V4_Copyright, V4_Tag, V4_Title, V4__AddLanguageToAll}
 
 import java.time.LocalDateTime
 import no.ndla.audioapi.{TestEnvironment, UnitSuite}
