+++ audio-api/src/test/scala/no/ndla/audioapi/db/migration/V8__MoveManuscriptFromPodcastToAudioTest.scala
@@ -5,9 +5,9 @@
  * See LICENSE
  */
 
-package audioapi.db.migration
+package no.ndla.audioapi.db.migration
 
-import audioapi.db.migration.V8__MoveManuscriptFromPodcastToAudio
+import no.ndla.audioapi.db.migration.V8__MoveManuscriptFromPodcastToAudio
 import no.ndla.audioapi.{TestEnvironment, UnitSuite}
 
 class V8__MoveManuscriptFromPodcastToAudioTest extends UnitSuite with TestEnvironment {
